requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
sqlparse==0.5.5
tzdata==2025.3
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==25.2
python-decouple==3.8
dj-database-url==2.1.0